Refactored Restaurants APIs
